+++ Cr-8/src/main/java/com/Cr_8/controllers/BookingController.java
@@ -33,8 +33,7 @@ import jakarta.validation.Valid;
 @RequestMapping("/api")
 @RestController
 public class BookingController {
-	@Autowired
-	private LabsService labsService ;
+
 	@Autowired
 	private BookingService bookService;
 	@Autowired
@@ -131,17 +130,6 @@ public class BookingController {
 		}
 		return new ResponseEntity<>("Invalid bookingId", HttpStatus.BAD_REQUEST);
 	}
-	@Tag(name = "Public Endpoint")
-	@GetMapping("/pub/labs")
-	public ResponseEntity<?> getAllLabs(){
-		List<Labs> allLabs = labsService.getAllLabs();
-		return new ResponseEntity<>(allLabs, HttpStatus.OK);
-	}
-	@Tag(name = "Dashboard Endpoint")
-	@PostMapping("/add-new-labs")
-	public ResponseEntity<?> addNewLabs(@RequestBody LabsDTO labsDTO){
-		String newlabs =labsService.addNewLabs(labsDTO.getName(), labsDTO.getDescription(),labsDTO.getScope(),labsDTO.getTargetDescription());
-		return new ResponseEntity<>(newlabs, HttpStatus.OK);
-	}
+	
 	
 }
